Split index.ts
